requirements.txt
MechanicalSoup==1.3.0
beautifulsoup4==4.12.3
python-magic==0.4.27
requests==2.31.0
PyYAML==6.0.1
ConfigArgParse==1.7
